AMLS-5648 | Removed old version of your_registration
