+++ app/views/status/your_registration.scala.html
@@ -17,8 +17,12 @@
 @import config.ApplicationConfig
 @import include._
 @import include.status._
+@import play.twirl.api.HtmlFormat
 
-@(regNo: String, businessName: Option[String], feeResponse: Option[FeeResponse], fromDuplicateSubmission: Boolean = false, yourRegistrationInfo: Html, canOrCannotTradeInformation: Html, unreadNotifications: Int = 0)(implicit request: Request[_],m:Messages, lang: Lang, appConfig: ApplicationConfig)
+@(regNo: String, businessName: Option[String], feeResponse: Option[FeeResponse] = None,
+        fromDuplicateSubmission: Boolean = false, yourRegistrationInfo: Html = HtmlFormat.empty,
+        unreadNotifications: Int = 0, registrationStatus: Html, feeInformation: Html, displayContactLink: Boolean = false,
+        withdrawOrDeregisterInformation: Html = HtmlFormat.empty)(implicit request: Request[_],m:Messages, lang: Lang, appConfig: ApplicationConfig)
 
 @main2(title = "Your registration", showURRBanner = true) {
 
@@ -55,10 +59,7 @@
         </div>
         <div id="registration-status" class="grid-layout__column grid-layout__column--1-3">
             <h2 class="heading-small">@Messages("your.registration.status")</h2>
-            <p>@Messages("your.registration.status.pending")
-                @canOrCannotTradeInformation
-            </p>
-
+            @registrationStatus
         </div>
         <div id="messages" class="grid-layout__column grid-layout__column--1-3">
             <h2 class="heading-small">@Messages("your.registration.messages")</h2>
@@ -68,27 +69,28 @@
 
     <div class="grid-layout margin-bottom-large">
         <div id="fees" class="grid-layout__column grid-layout__column--1-3">
-            <h2 class="heading-small">@Messages("status.fees")</h2>
-            <p>@Messages("your.registration.fees")
-                <br />
-                <a id="how-to-pay" href="how-to-pay">@Messages("status.howtopay")</a>
-            </p>
+            @feeInformation
         </div>
         <div class="grid-layout__column grid-layout__column--1-3"></div>
         <div class="grid-layout__column grid-layout__column--1-3"></div>
     </div>
 
-    <div>
-        <p>@Messages("status.submissionreadyforreview.description.5")
-            <a id="contact-link" href="https://www.gov.uk/government/organisations/hm-revenue-customs/contact/money-laundering">@Messages("status.submissionreadyforreview.description.5.link")</a>@Messages("full.stop")
-        </p>
-    </div>
+    @if(displayContactLink) {
+        <div>
+            <p>@Messages("status.submissionreadyforreview.description.5")
+                <a id="contact-link" href="https://www.gov.uk/government/organisations/hm-revenue-customs/contact/money-laundering">@Messages("status.submissionreadyforreview.description.5.link")</a>
+                @Messages("full.stop")
+            </p>
+        </div>
+    }
 
-    <div>
-        <p class="no-print">
-            @Messages("your.registration.withdraw")
-            <a id="withdraw-application" href="@controllers.withdrawal.routes.WithdrawApplicationController.get()">@Messages("status.withdraw.link-text")</a>@Messages("full.stop")
-        </p>
+    <div class="grid-layout">
+        <div class="grid-layout__column--2-3">
+            @withdrawOrDeregisterInformation
+        </div>
+        <div class="grid-layout__column grid-layout__column--1-3">
+            @* empty for layout reasons *@
+        </div>
     </div>
 
     @before_you_go()
